ref flag makes itself unique
The ref flag makes itself unique by appending 'a', 'b' etc.
mimicking Mendeley behaviour. This involves a new function
create_identifier (papis.utils) which takes a list or string (in this case
string.ascii_lowercase) and creates a list of possible strings as
it iterates over the list.

The ref flag then sets itself (papis.crossref), checks whether it
matches a ref field in a list of all ref fields. If it is already
in the list of references it appends itself with a string using
the create_identifier (papis.utils) command. If it still appears,
it repeats until it is unique.