Ensure `files` key is present in document
Sometimes when adding a document with no files to a library and then
later trying to use `addto` there would be an error because the document
(i.e. the YAML file) had no `files: []` entry. This commit ensures that
in the event that `files` key is not present that the key is added and
instantiated with an empty list.