OptionsList: create the pool on demand instead
This seems to have been causing some weird hangs on cpython 3.8 [1].
Current guess is that `__del__` wasn't called when expected anymore,
which the docs warn about [2], so the pool wasn't closed properly
and it deadlocked. Checked this with some of the other uses of
`multiprocessing` in the code and this is the only one that was
causing the hang.

[1] https://bugs.python.org/issue38744
[2] https://docs.python.org/3/reference/datamodel.html#object.__del__