tests/downloaders: rework retrieving and storing external and expected results
Basically adds two functions: `get_remote_resource` and
`get_local_resource`. Both take a filename from which to load cached
data and an url that can be used to update it from another resource.

Then, a `PAPIS_UPDATE_RESOURCES` environment variable controls if new
data is downloaded or existing cached data is checked. This should allow
* running the test without a network connection
* bulk updating resources to their latest versions.

This is especially useful for downloaders that have to
* work with the latest version of a website
* parse existing data correctly