flake: Override papis def in packageOverrides
We are going to be reorganizing the extras, and to ensure backwards
compatibility we're going to have a self-referential dependency

    papis[develop] -> papis[linting,testing,typing]

In order for pyproject.nix to be able to support this kind of
dependency, we need it to use our papis in its build inputs, rather than
the nixpkgs papis, which doesn't have these extras.
Thus, we add our papis to python's packageOverrides.

The reason this breaks is that the dependency resolution just looks for
what python.pkgs.papis is bound to, rather than selecting some package
satisfying version constraints out of a pool of possible packages, which
is in hindsight an unreasonable ask.

However, pyproject.nix has validators.validateVersionConstraints, which
appears to be able to at least catch this kind of problems? Testing and
adding it is left to future work.

References:
- [1][2][3] illustrate how currently in the nixpkgs ecosystem there is
  no expectation that nix will do version resolution

- [4][5] show how our implementation prior to this caused this bug

[1]: https://github.com/NixOS/nixpkgs/issues/93327
[2]: https://github.com/NixOS/nixpkgs/issues/9682#issuecomment-658424656
[3]: http://www.chriswarbo.net/projects/nixos/nix_dependencies.html

[4]: https://github.com/pyproject-nix/pyproject.nix/issues/278
[5]: https://github.com/hseg/papis/pull/1/files