Reworked the remote file logic in the addto command.
This fixes the regression, where addto with urls doesn't work, while
still working as expected with the `--link` option, i.e., downloaded
files are never symlinked (they are located in a tmp directory that is
likely cleared upon reboot). However, local files still respect the
`--link` option.

The cli parameters for addto are updated, such that -f now also takes
URL parameters. The split in `-f` for local files and `-u` was never
really useful and was just introduced, to be able to check for non-existing
files while doing the cli args parsing with click.

Co-authored-by: Alex Fikl <alexfikl@gmail.com>